feat: 岗位服务接口

+++ src/modules/system/repository/sys_user_post.go
@@ -5,7 +5,7 @@ import "mask_api_gin/src/modules/system/model"
 // ISysUserPost 用户与岗位关联表 数据层接口
 type ISysUserPost interface {
 	// CountUserPostByPostId 通过岗位ID查询岗位使用数量
-	CountUserPostByPostId(postId string) int
+	CountUserPostByPostId(postId string) int64
 
 	// BatchUserPost 批量新增用户岗位信息
 	BatchUserPost(sysUserPosts []model.SysUserPost) int64
